changed rectangle dimensions to double.

+++ src/cs3500/animator/view/SwingViewPanel.java
@@ -42,7 +42,8 @@ class SwingViewPanel extends JPanel {
         Rectangle r = (Rectangle) d;
         g2.setColor(new Color(
                 (int) r.getColor().getR(), (int) r.getColor().getG(), (int) r.getColor().getB()));
-        g2.drawRect((int) r.getPos().getX(), (int) r.getPos().getY(), r.getWidth(), r.getHeight());
+        g2.drawRect((int) r.getPos().getX(), (int) r.getPos().getY(),
+                (int) r.getWidth(), (int) r.getHeight());
       }
     }
